Rename mutationadapter module to controller module
